JFG-890: create DefaultHttpInvoker
